requirements-api.txt
# API service dependencies (Render/Railway/Fly)
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0

# Inference/runtime
pandas>=2.2,<2.4
numpy>=2.1,<2.4
scikit-learn>=1.6,<1.8
joblib>=1.4,<2.0
pyarrow>=18,<20

# Data + logging
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0
loguru>=0.7,<1.0
plotly>=5.22,<6.0
